refactor: move all the diagnostics related code to its own module exposing minimal API to be used in our custom "middleware"
